Remove tensor conversion utils make it local in model to remove with legacy evaluate
